Generate published API entry point from src/api.ts
The npm package `@vscode/python-environments` previously duplicated the
entire API definition in `api/src/main.ts`, a near-copy of `src/api.ts`
that was hand-maintained and had drifted.

Make `src/api.ts` the single source of truth by folding the publish-only
runtime helper (`PythonEnvironments.api()` and `EXTENSION_ID`) into it, and
stop committing `api/src/main.ts`. The publish pipeline now copies
`src/api.ts` to `api/src/main.ts` before compiling the package.

- src/api.ts: add the `PythonEnvironments.api()` helper and `EXTENSION_ID`
- api/src/main.ts: remove (now generated); ignore via api/.gitignore
- build/azure-pipeline.npm.yml: copy src/api.ts before compile
- CONTRIBUTING.md: document the copy model
